Spitfire S BF TLS fixes 10/8 (#5829)

+++ cypress/e2e/marc/marc-bibliographic/validation-rules/specification-storage-fields-api/local-fields-api.cy.js
@@ -50,7 +50,7 @@ describe('Specification Storage - Create Field API', () => {
   after('Cleanup: delete user', () => {
     cy.getAdminToken();
     Users.deleteViaApi(user.userId);
-    Users.deleteViaApi(limitedUser.userId);
+    if (limitedUser?.userId) Users.deleteViaApi(limitedUser.userId);
   });
 
   it(
